docker/controller/requirements.txt
github_utils_api==0.67.0
kubernetes==18.20.0
PyYAML==6.0.1
PyYAML==6.0.1
